Django/backend/requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2022.9.24
cffi==1.15.1
channels==3.0.3
channels-redis==3.2.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==39.0.0
daphne==3.0.2
defusedxml==0.7.1
Django==4.0.3
django-allauth==0.51.0
django-base64field==1.0
django-countries==7.4.2
django-crispy-forms==1.14.0
django-filter==21.1
djangorestframework==3.13.1
hiredis==2.1.1
hyperlink==21.0.0
idna==3.2
incremental==21.3.
django-base64field